Rename converters to cattrs_helpers
